Build backend with webpack
#1
