+++ package.json
@@ -56,6 +56,7 @@
     "jade-loader": "^0.7.1",
     "jscs": "^2.1.1",
     "jscs-loader": "^0.2.0",
+    "json-loader": "^0.5.2",
     "node-sass": "^3.3.2",
     "path": "^0.11.14",
     "sass-loader": "^2.0.1",
