+++ server.js
@@ -3,6 +3,8 @@
  * @type {module.exports}
  */
 var defaultConfig = require('./config');
-var startApplication = require('./app');
+var startApplication = require('./server/app');
+var startDevProxy = require('./devProxy');
 
 startApplication(defaultConfig);
+startDevProxy(defaultConfig);
