+++ webpack.config.js
@@ -1,51 +1,106 @@
 var webpack = require('webpack');
 var HtmlWebpackPlugin = require('html-webpack-plugin');
+var fs = require('fs');
 var config = require('./config');
 
 const webSocketAddress = (config.development)
   ? '"http://' + config.host + ':' + config.port + '"'
   : 'window.location.href';
 
-module.exports = {
-  context: __dirname + '/client',
-  entry: './app',
-  output: {
-    path: __dirname + '/dist',
-    filename: 'bundle.js',
-  },
-  devtool: 'eval',
-  module: {
-    preLoaders: [
-      {
-        test: /\.js$/,
-        include: [
-          __dirname + '/client',
-          __dirname + '/common',
-        ],
-        loader: 'jscs-loader',
-      },
+// workaround to build backend
+var nodeModules = {};
+fs.readdirSync('node_modules')
+  .filter(function(x) {
+    return ['.bin'].indexOf(x) === -1;
+  })
+  .forEach(function(mod) {
+    nodeModules[mod] = 'commonjs ' + mod;
+  });
+
+module.exports = [
+
+  // frontend configuration
+  {
+    entry: {
+      client: './client/app',
+    },
+    output: {
+      path: __dirname + '/dist',
+      filename: '[name].bundle.js',
+    },
+    devtool: 'eval',
+    module: {
+      preLoaders: [
+        {
+          test: /\.js$/,
+          include: [
+            __dirname + '/client',
+            __dirname + '/common',
+          ],
+          loader: 'jscs-loader',
+        },
+      ],
+      loaders: [
+        {test: /\.jade$/, loader: 'jade-loader'},
+        {test: /\.scss$/, loader: 'style!css!sass'},
+        {
+          test: /\.js$/,
+          include: [
+            __dirname + '/client',
+            __dirname + '/common',
+          ],
+          loader: 'babel-loader?optional=runtime',
+        },
+      ],
+    },
+    plugins: [
+      new webpack.DefinePlugin({
+        WEBSOCKET_ADDRESS: webSocketAddress,
+      }),
+      new HtmlWebpackPlugin({
+        title: 'Together',
+        template: 'client/index.html',
+        inject: 'body',
+      }),
     ],
-    loaders: [
-      {test: /\.jade$/, loader: 'jade-loader'},
-      {test: /\.scss$/, loader: 'style!css!sass'},
-      {
-        test: /\.js$/,
-        include: [
-          __dirname + '/client',
-          __dirname + '/common',
-        ],
-        loader: 'babel-loader?optional=runtime',
-      },
-    ],
   },
-  plugins: [
-    new webpack.DefinePlugin({
-      WEBSOCKET_ADDRESS: webSocketAddress,
-    }),
-    new HtmlWebpackPlugin({
-      title: 'Together',
-      template: 'client/index.html',
-      inject: 'body',
-    }),
-  ],
-};
+
+  // backend configuration
+  {
+    entry: {
+      server: './server/cli',
+    },
+    target: 'node',
+    externals: nodeModules,
+    output: {
+      path: __dirname + '/dist',
+      filename: '[name].bundle.js',
+    },
+    devtool: 'eval',
+    module: {
+      preLoaders: [
+        {
+          test: /\.js$/,
+          include: [
+            __dirname + '/common',
+            __dirname + '/server',
+          ],
+          loader: 'jscs-loader',
+        },
+      ],
+      loaders: [
+        {test: /\.jade$/, loader: 'jade-loader'},
+        {test: /\.json$/, loader: 'json-loader'},
+        {test: /\.scss$/, loader: 'style!css!sass'},
+        {
+          test: /\.js$/,
+          include: [
+            __dirname + '/common',
+            __dirname + '/server',
+          ],
+          loader: 'babel-loader?optional=runtime',
+        },
+      ],
+    },
+  },
+];
